fix audio pacing chunk duration type mismatch from float sample rate

+++ linux/src/main/kotlin/com/osuradio/app/audio/DesktopPlayer.kt
@@ -420,7 +420,7 @@ class DesktopPlayer(
                 continue
             }
 
-            val chunkNs = framesToWrite.toLong() * 1000000000L / sampleRate
+            val chunkNs = framesToWrite.toLong() * 1000000000L / sampleRate.toLong()
             val t0 = System.nanoTime()
             writeToLine(chunk, framesToWrite)
             val remainingNs = chunkNs - (System.nanoTime() - t0)
